dot_config/agents/skills/gmail-improved/pyproject.toml
[project]
name = "gmail-improved"
version = "0.1.0"
description = "Gmail CLI for Claude Code skill"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "pyyaml>=6.0",
]
